Optimized structuring of JsonGauge, changed method naming in JsonGaugeInternal (using keyword fits instead of matches)
